requirements.txt
fastapi
uvicorn
google-generativeai
faiss-cpu
python-dotenv
requests
beautifulsoup4
pandas
numpy